gen: code by fern

+++ src/dify_sdk/types/upload_file.py
@@ -47,7 +47,7 @@ class UploadFile(UniversalBaseModel):
     Creator ID
     """
 
-    created_at: typing.Optional[int] = pydantic.Field(default=None)
+    created_at: typing.Optional[float] = pydantic.Field(default=None)
     """
     Creation timestamp
     """
